+++ tests/integration/run_all.sh
@@ -114,6 +114,7 @@ ALL_CASES=(
     test_smp
     test_smp_tss
     test_smp_init_order
+    test_fpu_smp
     test_gui
     test_gui_usb
     test_gbrowser